zmarkdown: extract image-to-figure
