+++ packages/zmarkdown/index.js
@@ -28,6 +28,7 @@ const remarkPing = require('remark-ping/src')
 const remarkSubSuper = require('remark-sub-super/src')
 const remarkTextr = require('./plugins/remark-textr')
 const remarkTrailingSpaceHeading = require('remark-heading-trailing-spaces')
+const remarkImageToFigure = require('./plugins/remark-image-to-figure')
 
 const remark2rehype = require('remark-rehype')
 
@@ -69,32 +70,6 @@ const wrappers = {
   ],
 }
 
-const imageToFigure = (img, index, parent) => {
-  if (parent.children.length === 1 && parent.type === 'paragraph') {
-    if (!img.alt) {
-      return
-    }
-    const figureCaptionNode = {
-      type: 'figcaption',
-      children: [
-        {
-          type: 'text',
-          value: img.alt,
-        },
-      ],
-      data: {
-        hName: 'figcaption',
-      },
-    }
-    parent.type = 'figure'
-    parent.children = [clone(img), figureCaptionNode]
-    parent.data = {
-      hName: 'figure',
-    }
-  }
-}
-
-
 const zmdParser = (config, target) => {
   const mdProcessor = unified()
     .use(remarkParse, config.reParse)
@@ -123,9 +98,8 @@ const zmdParser = (config, target) => {
     .use(remarkPing, config.ping)
     .use(remarkSubSuper)
     .use(remarkTrailingSpaceHeading)
+    .use(remarkImageToFigure)
     .use(() => (tree, vfile) => {
-      // implement old functionality that transforms block images into figure
-      visit(tree, 'image', imageToFigure)
       /* extract some metadata for frontends */
 
       // if we don't have any headings, we add a flag to disable
